refactor: 멤버 서비스 패키지 변경
